+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * [#1488](https://github.com/crypto-org-chain/cronos/pull/1488) Enable optimistic execution.
 * [#1490](https://github.com/crypto-org-chain/cronos/pull/1490) Update cometbft to v0.38.8.
 * [#1491](https://github.com/crypto-org-chain/cronos/pull/1491) Free slice data in HasAtVersion.
+* [#1498](https://github.com/crypto-org-chain/cronos/pull/1498) Check validity with extractSliceBytes to avoid copying slice data.
 
 ### Bug Fixes